formatting

+++ src/BadgeStorage.sol
@@ -21,20 +21,20 @@ contract BadgeStorage is ReentrancyGuard {
     }
 
     struct StorageItem {
-        uint itemId;
+        uint256 itemId;
         address nftContract;
-        uint tokenId;
+        uint256 tokenId;
         address payable creator; //NFT creator
         address payable owner; //owner after it's being transfered
         bool claim;
     }
 
-    mapping(uint => StorageItem) private idToStorageItem;
+    mapping(uint256 => StorageItem) private idToStorageItem;
 
     event ItemStored(
-        uint indexed itemId,
+        uint256 indexed itemId,
         address indexed nftContract,
-        uint indexed tokenId,
+        uint256 indexed tokenId,
         address creator,
         address owner,
         bool claim
@@ -42,42 +42,20 @@ contract BadgeStorage is ReentrancyGuard {
 
     /* Stores an NFT on the storage before distributed */
     // create market item
-    function storeItem(
-        address nftContract,
-        uint tokenId
-    ) public payable nonReentrant {
-        uint itemId = ++itemIds;
-        require(
-            idToStorageItem[itemId].claim == false,
-            "You can't store a redeemed token!"
-        );
-        idToStorageItem[itemId] = StorageItem(
-            itemId,
-            nftContract,
-            tokenId,
-            payable(msg.sender),
-            payable(address(0)),
-            false
-        );
+    function storeItem(address nftContract, uint256 tokenId) public payable nonReentrant {
+        uint256 itemId = ++itemIds;
+        require(idToStorageItem[itemId].claim == false, "You can't store a redeemed token!");
+        idToStorageItem[itemId] =
+            StorageItem(itemId, nftContract, tokenId, payable(msg.sender), payable(address(0)), false);
         IERC721(nftContract).transferFrom(msg.sender, address(this), tokenId);
 
-        emit ItemStored(
-            itemId,
-            nftContract,
-            tokenId,
-            msg.sender,
-            address(0),
-            false
-        );
+        emit ItemStored(itemId, nftContract, tokenId, msg.sender, address(0), false);
     }
 
     /* Transfers the NFT's ownership to attendee (msg.sender)*/
     // buy nft
-    function claimItem(
-        address nftContract,
-        uint itemId
-    ) public payable nonReentrant {
-        uint tokenId = idToStorageItem[itemId].tokenId;
+    function claimItem(address nftContract, uint256 itemId) public payable nonReentrant {
+        uint256 tokenId = idToStorageItem[itemId].tokenId;
 
         IERC721(nftContract).transferFrom(address(this), msg.sender, tokenId);
         idToStorageItem[itemId].owner = payable(msg.sender);
@@ -87,20 +65,16 @@ contract BadgeStorage is ReentrancyGuard {
     }
 
     /*Returns all NFTs which has not been claimed*/
-    function fetchStorageItems()
-        public
-        view
-        returns (StorageItem[] memory items)
-    {
-        uint itemCount = itemIds;
-        uint notRedeemedItemCount = itemIds - itemsRedeemed;
-        uint currentIndex = 0;
+    function fetchStorageItems() public view returns (StorageItem[] memory items) {
+        uint256 itemCount = itemIds;
+        uint256 notRedeemedItemCount = itemIds - itemsRedeemed;
+        uint256 currentIndex = 0;
 
         StorageItem[] memory items = new StorageItem[](notRedeemedItemCount);
 
-        for (uint i = 0; i < itemCount; i++) {
+        for (uint256 i = 0; i < itemCount; i++) {
             if (idToStorageItem[i + 1].owner == address(0)) {
-                uint currentId = 1 + i;
+                uint256 currentId = 1 + i;
                 StorageItem storage currentItem = idToStorageItem[currentId];
 
                 items[currentIndex] = currentItem;
@@ -112,20 +86,20 @@ contract BadgeStorage is ReentrancyGuard {
 
     /*Returns all NFTs which has been redeemed*/
     function fetchMyNFTs() public view returns (StorageItem[] memory items) {
-        uint totalItemCount = itemIds;
-        uint itemCount = 0;
-        uint currentIndex = 0;
+        uint256 totalItemCount = itemIds;
+        uint256 itemCount = 0;
+        uint256 currentIndex = 0;
 
-        for (uint i = 0; i < totalItemCount; i++) {
+        for (uint256 i = 0; i < totalItemCount; i++) {
             if (idToStorageItem[i + 1].owner == msg.sender) {
                 itemCount = i + 1;
             }
         }
 
         StorageItem[] memory items = new StorageItem[](itemCount);
-        for (uint i = 0; i < totalItemCount; i++) {
+        for (uint256 i = 0; i < totalItemCount; i++) {
             if (idToStorageItem[i + 1].owner == msg.sender) {
-                uint currentId = i + 1;
+                uint256 currentId = i + 1;
                 StorageItem storage currentItem = idToStorageItem[currentId];
                 items[currentIndex] = currentItem;
                 currentIndex += 1;
@@ -135,16 +109,12 @@ contract BadgeStorage is ReentrancyGuard {
     }
 
     // returns only NFTs the creator has created
-    function fetchItemsCreated()
-        public
-        view
-        returns (StorageItem[] memory items)
-    {
-        uint totalItemCount = itemIds;
-        uint itemCount = 0;
-        uint currentIndex = 0;
+    function fetchItemsCreated() public view returns (StorageItem[] memory items) {
+        uint256 totalItemCount = itemIds;
+        uint256 itemCount = 0;
+        uint256 currentIndex = 0;
 
-        for (uint i = 0; i < totalItemCount; i++) {
+        for (uint256 i = 0; i < totalItemCount; i++) {
             if (idToStorageItem[i + 1].creator == msg.sender) {
                 itemCount += 1;
             }
@@ -152,9 +122,9 @@ contract BadgeStorage is ReentrancyGuard {
 
         StorageItem[] memory items = new StorageItem[](itemCount);
 
-        for (uint i = 0; i < totalItemCount; i++) {
+        for (uint256 i = 0; i < totalItemCount; i++) {
             if (idToStorageItem[i + 1].creator == msg.sender) {
-                uint currentId = i + 1;
+                uint256 currentId = i + 1;
                 StorageItem storage currentItem = idToStorageItem[currentId];
                 items[currentIndex] = currentItem;
                 currentIndex += 1;
